requirements-dev.txt
polygon-api-client
pymongo
certifi
alpaca-py
yfinance
scipy
pandas
cython
fastapi
uvicorn
motor
python-dotenv
stats
PyWavelets
ripser
datetime
ta-lib
pandas-ta
python-dotenv==1.0.1
black==24.10.0
isort==5.13.2